# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EMI Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Shop {
  id: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  createdAt?: string;
}

export interface ShopInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
}

export interface Customer {
  id: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  createdAt?: string;
}

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
}

export interface Product {
  id: number;
  name: string;
  shopId: number;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  shopName?: string | null;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  shopId: number;
  /** @nullable */
  price?: number | null;
}

export interface EmiOrder {
  id: number;
  customerId: number;
  shopId: number;
  /** @nullable */
  productId?: number | null;
  productName: string;
  totalPrice: number;
  downPayment: number;
  emiMonths: number;
  monthlyAmount: number;
  totalPaid?: number;
  remainingAmount?: number;
  status: string;
  purchaseDate: string;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  shopName?: string | null;
}

export interface EmiOrderInput {
  customerId: number;
  shopId: number;
  /** @nullable */
  productId?: number | null;
  /** @minLength 1 */
  productName: string;
  totalPrice: number;
  downPayment: number;
  emiMonths: number;
  monthlyAmount: number;
  purchaseDate: string;
}

export interface EmiOrderUpdate {
  productName?: string;
  totalPrice?: number;
  downPayment?: number;
  emiMonths?: number;
  monthlyAmount?: number;
  status?: string;
  purchaseDate?: string;
}

export interface EmiPayment {
  id: number;
  emiOrderId: number;
  amount: number;
  paymentDate: string;
  paymentMethod: string;
  /** @nullable */
  notes?: string | null;
}

export interface EmiOrderDetail {
  id: number;
  customerId: number;
  shopId: number;
  /** @nullable */
  productId?: number | null;
  productName: string;
  totalPrice: number;
  downPayment: number;
  emiMonths: number;
  monthlyAmount: number;
  totalPaid?: number;
  remainingAmount?: number;
  status: string;
  purchaseDate: string;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  shopName?: string | null;
  payments: EmiPayment[];
}

export interface EmiPaymentInput {
  amount: number;
  paymentDate: string;
  paymentMethod: string;
  /** @nullable */
  notes?: string | null;
}

export interface DashboardSummary {
  totalActiveOrders: number;
  totalCompletedOrders?: number;
  totalOrders: number;
  totalDueAmount: number;
  totalPaidAmount: number;
  overdueOrders: number;
  thisMonthCollected?: number;
}

export interface ShopStat {
  shopId: number;
  shopName: string;
  totalOrders: number;
  totalDue: number;
  totalPaid: number;
}

export type ListProductsParams = {
/**
 * @nullable
 */
shopId?: number | null;
};

export type ListEmiOrdersParams = {
/**
 * @nullable
 */
shopId?: number | null;
/**
 * @nullable
 */
customerId?: number | null;
/**
 * @nullable
 */
status?: string | null;
};