# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EMI Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: shops
    description: Shop/Showroom management
  - name: customers
    description: Customer management
  - name: products
    description: Product management
  - name: emi-orders
    description: EMI order management
  - name: payments
    description: EMI payment management
  - name: dashboard
    description: Dashboard summary

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /shops:
    get:
      operationId: listShops
      tags: [shops]
      summary: List all shops
      responses:
        "200":
          description: List of shops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shop"
    post:
      operationId: createShop
      tags: [shops]
      summary: Create a new shop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"

  /shops/{id}:
    get:
      operationId: getShop
      tags: [shops]
      summary: Get shop by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "404":
          description: Not found
    put:
      operationId: updateShop
      tags: [shops]
      summary: Update shop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
    delete:
      operationId: deleteShop
      tags: [shops]
      summary: Delete shop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: shopId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /emi-orders:
    get:
      operationId: listEmiOrders
      tags: [emi-orders]
      summary: List all EMI orders
      parameters:
        - name: shopId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: customerId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of EMI orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmiOrder"
    post:
      operationId: createEmiOrder
      tags: [emi-orders]
      summary: Create a new EMI order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmiOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmiOrder"

  /emi-orders/{id}:
    get:
      operationId: getEmiOrder
      tags: [emi-orders]
      summary: Get EMI order by ID with payment history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: EMI order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmiOrderDetail"
        "404":
          description: Not found
    put:
      operationId: updateEmiOrder
      tags: [emi-orders]
      summary: Update EMI order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmiOrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmiOrder"
    delete:
      operationId: deleteEmiOrder
      tags: [emi-orders]
      summary: Delete EMI order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /emi-orders/{id}/payments:
    get:
      operationId: listEmiPayments
      tags: [payments]
      summary: List payments for an EMI order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmiPayment"
    post:
      operationId: createEmiPayment
      tags: [payments]
      summary: Record a payment for an EMI order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmiPaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmiPayment"

  /payments/{paymentId}:
    delete:
      operationId: deleteEmiPayment
      tags: [payments]
      summary: Delete a payment record
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/due-this-month:
    get:
      operationId: getDueThisMonth
      tags: [dashboard]
      summary: Get EMI orders with payments due this month
      responses:
        "200":
          description: Due this month
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmiOrder"

  /dashboard/shop-stats:
    get:
      operationId: getShopStats
      tags: [dashboard]
      summary: Get per-shop EMI statistics
      responses:
        "200":
          description: Shop stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShopStat"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Shop:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string

    ShopInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]

    Customer:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string

    CustomerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]

    Product:
      type: object
      required: [id, name, shopId]
      properties:
        id:
          type: integer
        name:
          type: string
        shopId:
          type: integer
        price:
          type: ["number", "null"]
        shopName:
          type: ["string", "null"]

    ProductInput:
      type: object
      required: [name, shopId]
      properties:
        name:
          type: string
          minLength: 1
        shopId:
          type: integer
        price:
          type: ["number", "null"]

    EmiOrder:
      type: object
      required: [id, customerId, shopId, productName, totalPrice, downPayment, emiMonths, monthlyAmount, status, purchaseDate]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        shopId:
          type: integer
        productId:
          type: ["integer", "null"]
        productName:
          type: string
        totalPrice:
          type: number
        downPayment:
          type: number
        emiMonths:
          type: integer
        monthlyAmount:
          type: number
        totalPaid:
          type: number
        remainingAmount:
          type: number
        status:
          type: string
        purchaseDate:
          type: string
        customerName:
          type: ["string", "null"]
        shopName:
          type: ["string", "null"]

    EmiOrderInput:
      type: object
      required: [customerId, shopId, productName, totalPrice, downPayment, emiMonths, monthlyAmount, purchaseDate]
      properties:
        customerId:
          type: integer
        shopId:
          type: integer
        productId:
          type: ["integer", "null"]
        productName:
          type: string
          minLength: 1
        totalPrice:
          type: number
        downPayment:
          type: number
        emiMonths:
          type: integer
        monthlyAmount:
          type: number
        purchaseDate:
          type: string

    EmiOrderUpdate:
      type: object
      properties:
        productName:
          type: string
        totalPrice:
          type: number
        downPayment:
          type: number
        emiMonths:
          type: integer
        monthlyAmount:
          type: number
        status:
          type: string
        purchaseDate:
          type: string

    EmiOrderDetail:
      type: object
      required: [id, customerId, shopId, productName, totalPrice, downPayment, emiMonths, monthlyAmount, status, purchaseDate, payments]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        shopId:
          type: integer
        productId:
          type: ["integer", "null"]
        productName:
          type: string
        totalPrice:
          type: number
        downPayment:
          type: number
        emiMonths:
          type: integer
        monthlyAmount:
          type: number
        totalPaid:
          type: number
        remainingAmount:
          type: number
        status:
          type: string
        purchaseDate:
          type: string
        customerName:
          type: ["string", "null"]
        shopName:
          type: ["string", "null"]
        payments:
          type: array
          items:
            $ref: "#/components/schemas/EmiPayment"

    EmiPayment:
      type: object
      required: [id, emiOrderId, amount, paymentDate, paymentMethod]
      properties:
        id:
          type: integer
        emiOrderId:
          type: integer
        amount:
          type: number
        paymentDate:
          type: string
        paymentMethod:
          type: string
        notes:
          type: ["string", "null"]

    EmiPaymentInput:
      type: object
      required: [amount, paymentDate, paymentMethod]
      properties:
        amount:
          type: number
        paymentDate:
          type: string
        paymentMethod:
          type: string
        notes:
          type: ["string", "null"]

    DashboardSummary:
      type: object
      required: [totalActiveOrders, totalDueAmount, totalPaidAmount, totalOrders, overdueOrders]
      properties:
        totalActiveOrders:
          type: integer
        totalCompletedOrders:
          type: integer
        totalOrders:
          type: integer
        totalDueAmount:
          type: number
        totalPaidAmount:
          type: number
        overdueOrders:
          type: integer
        thisMonthCollected:
          type: number

    ShopStat:
      type: object
      required: [shopId, shopName, totalOrders, totalDue, totalPaid]
      properties:
        shopId:
          type: integer
        shopName:
          type: string
        totalOrders:
          type: integer
        totalDue:
          type: number
        totalPaid:
          type: number