# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EMI Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  id: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  createdAt?: string;
}