# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EMI Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalActiveOrders: number;
  totalCompletedOrders?: number;
  totalOrders: number;
  totalDueAmount: number;
  totalPaidAmount: number;
  overdueOrders: number;
  thisMonthCollected?: number;
}