# lib/api-zod/src/generated/types/emiOrderDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EMI Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { EmiPayment } from './emiPayment';

export interface EmiOrderDetail {
  id: number;
  customerId: number;
  shopId: number;
  /** @nullable */
  productId?: number | null;
  productName: string;
  totalPrice: number;
  downPayment: number;
  emiMonths: number;
  monthlyAmount: number;
  totalPaid?: number;
  remainingAmount?: number;
  status: string;
  purchaseDate: string;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  shopName?: string | null;
  payments: EmiPayment[];
}